controller input; box2d pixels->meters; dashing

+++ core/src/galenscovell/soulslite/actors/components/SizeComponent.scala
@@ -3,6 +3,6 @@ package galenscovell.soulslite.actors.components
 import com.badlogic.ashley.core.Component
 
 
-class SizeComponent(val size: Int) extends Component {
+class SizeComponent(val size: Float) extends Component {
 
 }
